phase_6_ultra_fast/backend/requirements.txt
fastapi
uvicorn[standard]
pandas
thefuzz
python-Levenshtein
python-dotenv
google-genai
pydantic
pytest
httpx
